Checking in BioFood
BioFood Read Me

This directory contains the code for BioFood.  The code
for the iPhone app and for the AppEngine app are in separate
directories.  The AppEngine app contains code for Loosestrife, another
(related) project that we worked on, in addition to the code for
BioFood.

Some important notes:
1) BioFood has, at various points in its
development, been called FlowerPower and GreenTouch.  So these legacy
names might still appear in the code, but they're referring to the same
project.

2) In the AppEngine code, the main.py file is what assigns
URLs to classes, and the classes have their GET and POST methods
exectued depending on the type of web request sent.

3) Pushing
AppEngine code to production requires you to be added through Google's
system, which we haven't done yet.  If you sent us (Michelle's writing
this, but Consuelo or Casey would likely also be able to do it) an email
with the specific accounts that you want added, we can take care of
it.

4) The code isn't pretty... our lab's focus is always on getting a
working prototype, so we never prioritize refactoring or code cleanup.
If you have any questions about specific things you see, I would be
happy to help, and I also know the entire code base, which is a plus.

-Kelsey Tempel